searchproduct

+++ server/graph/model/models_gen.go
@@ -76,12 +76,12 @@ type NewWishlist struct {
 }
 
 type SearchProduct struct {
-	Keyword        *string `json:"keyword"`
-	MinPrice       *int    `json:"minPrice"`
-	MaxPrice       *int    `json:"maxPrice"`
-	OrderBy        *string `json:"orderBy"`
-	CategoryID     *string `json:"categoryID"`
-	IsDiscount     *bool   `json:"isDiscount"`
-	CreatedAtRange *int    `json:"createdAtRange"`
-	HighRating     *bool   `json:"highRating"`
+	Keyword        *string  `json:"keyword"`
+	MinPrice       *float64 `json:"minPrice"`
+	MaxPrice       *float64 `json:"maxPrice"`
+	OrderBy        *string  `json:"orderBy"`
+	CategoryID     *string  `json:"categoryID"`
+	IsDiscount     *bool    `json:"isDiscount"`
+	CreatedAtRange *int     `json:"createdAtRange"`
+	HighRating     *bool    `json:"highRating"`
 }
